Revert "Moved cosmos query test to azure cosmos module"
This reverts commit bdd349d0f21d92d93b19faa717443e6167c91e7a.
