+++ data-model/src/main/java/io/micronaut/data/model/query/builder/sql/SqlQueryBuilder.java
@@ -250,7 +250,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
             String joinTableName = associationMetadata
                     .stringValue(ANN_JOIN_TABLE, "name")
                     .orElseGet(() ->
-                            getMappedName(namingStrategy, association)
+                            namingStrategy.mappedName(association)
                     );
             dropStatements.add("DROP TABLE " + (escape ? quote(joinTableName) : joinTableName) + ";");
         }
@@ -278,7 +278,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
             String joinTableName = annotationMetadata
                     .stringValue(ANN_JOIN_TABLE, "name")
                     .orElseGet(() ->
-                            getMappedName(namingStrategy, association)
+                            namingStrategy.mappedName(association)
                     );
             List<String> leftJoinColumns = resolveJoinTableJoinColumns(annotationMetadata, true, entity, namingStrategy);
             List<String> rightJoinColumns = resolveJoinTableJoinColumns(annotationMetadata, false, association.getAssociatedEntity(), namingStrategy);
@@ -344,7 +344,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
                 String joinTableName = annotationMetadata
                         .stringValue(ANN_JOIN_TABLE, "name")
                         .orElseGet(() ->
-                                getMappedName(namingStrategy, association)
+                                namingStrategy.mappedName(association)
                         );
                 if (escape) {
                     joinTableName = quote(joinTableName);
@@ -431,7 +431,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
             }
             boolean finalGeneratePkAfterColumns = generatePkAfterColumns;
             for (PersistentPropertyPath pp : ids) {
-                String column = getMappedName(namingStrategy, pp.getAssociations(), pp.getProperty());
+                String column = namingStrategy.mappedName(pp.getAssociations(), pp.getProperty());
                 if (escape) {
                     column = quote(column);
                 }
@@ -446,7 +446,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
         }
         PersistentProperty version = entity.getVersion();
         if (version != null) {
-            String column = getMappedName(namingStrategy, Collections.emptyList(), version);
+            String column = namingStrategy.mappedName(Collections.emptyList(), version);
             if (escape) {
                 column = quote(column);
             }
@@ -455,7 +455,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
         }
 
         BiConsumer<List<Association>, PersistentProperty> addColumn = (associations, property) -> {
-            String column = getMappedName(namingStrategy, associations, property);
+            String column = namingStrategy.mappedName(associations, property);
             if (escape) {
                 column = quote(column);
             }
@@ -713,7 +713,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
         }
         List<String> columns = new ArrayList<>();
         traversePersistentProperties(identity, (associations, property) -> {
-            String columnName = getMappedName(namingStrategy, associations, property);
+            String columnName = namingStrategy.mappedName(associations, property);
             columns.add(columnName);
         });
         return columns;
@@ -793,7 +793,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
                     traversePersistentProperties(associatedEntity, includeIdentity, true, (propertyAssociations, prop) -> {
                         String columnName;
                         if (computePropertyPaths()) {
-                            columnName = getMappedName(namingStrategy, propertyAssociations, prop);
+                            columnName = namingStrategy.mappedName(propertyAssociations, prop);
                         } else {
                             columnName = asPath(propertyAssociations, prop);
                         }
@@ -840,7 +840,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
             if (transformed != null) {
                 sb.append(transformed).append(AS_CLAUSE).append(useAlias ? columnAlias : property.getPersistedName());
             } else {
-                String column = getMappedName(namingStrategy, associations, property);
+                String column = namingStrategy.mappedName(associations, property);
                 column = escapeColumnIfNeeded(column, escape);
                 sb.append(alias).append(DOT).append(column);
                 if (useAlias) {
@@ -964,7 +964,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
                         }
                     });
 
-                    String columnName = getMappedName(namingStrategy, associations, property);
+                    String columnName = namingStrategy.mappedName(associations, property);
                     if (escape) {
                         columnName = quote(columnName);
                     }
@@ -994,7 +994,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
                 }
             });
 
-            String columnName = getMappedName(namingStrategy, Collections.emptyList(), version);
+            String columnName = namingStrategy.mappedName(Collections.emptyList(), version);
             if (escape) {
                 columnName = quote(columnName);
             }
@@ -1046,7 +1046,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
 
                 }
 
-                String columnName = getMappedName(namingStrategy, associations, property);
+                String columnName = namingStrategy.mappedName(associations, property);
                 if (escape) {
                     columnName = quote(columnName);
                 }
@@ -1348,7 +1348,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
 
             String joinTableName = annotationMetadata
                     .stringValue(ANN_JOIN_TABLE, "name")
-                    .orElseGet(() -> getMappedName(namingStrategy, association));
+                    .orElseGet(() -> namingStrategy.mappedName(association));
             String joinTableAlias = annotationMetadata
                     .stringValue(ANN_JOIN_TABLE, "alias")
                     .orElseGet(() -> currentJoinAlias + joinTableName + "_");
@@ -1458,7 +1458,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
         }
         if (onLeftColumns.isEmpty()) {
             traversePersistentProperties(leftProperty, (associations, p) -> {
-                String column = getMappedName(getNamingStrategy(leftProperty.getOwner()), merge(leftPropertyAssociations, associations), p);
+                String column = getNamingStrategy(leftProperty.getOwner()).mappedName(merge(leftPropertyAssociations, associations), p);
                 onLeftColumns.add(column);
             });
             if (onLeftColumns.isEmpty()) {
@@ -1467,7 +1467,7 @@ public class SqlQueryBuilder extends AbstractSqlLikeQueryBuilder implements Quer
         }
         if (onRightColumns.isEmpty()) {
             traversePersistentProperties(rightProperty, (associations, p) -> {
-                String column = getMappedName(getNamingStrategy(rightProperty.getOwner()), merge(rightPropertyAssociations, associations), p);
+                String column = getNamingStrategy(rightProperty.getOwner()).mappedName(merge(rightPropertyAssociations, associations), p);
                 onRightColumns.add(column);
             });
         }
